server/requirements.txt
# ============================================
# FastAPI 核心依赖
# ============================================
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.9

# ============================================
# 数据验证和配置
# ============================================
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0

# ============================================
# 数据库相关
# ============================================
sqlalchemy>=2.0.36
aiomysql>=0.2.0
pymysql>=1.1.0
alembic>=1.14.0

# ============================================
# Redis 缓存
# ============================================
redis>=5.2.0

# ============================================
# 认证和安全
# ============================================
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.2.0

# ============================================
# 其他工具
# ============================================
# HTTP 客户端(用于微信 API 调用)
httpx>=0.28.0
